requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.110.1
fastapi-responses==0.2.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
psycopg2-binary
pydantic==2.7.0
pydantic-core==2.18.1
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing-extensions==4.11.0
uvicorn==0.29.0
python-dotenv==0.21.0